ilogic/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-crispy-forms==2.0
django-filter==23.2
django-import-export==3.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
django-single-session==0.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
numpy==1.25.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.1.0
Pillow==10.0.0
pycryptodome==3.18.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
urllib3==2.0.4
xlrd==2.0.1
xlwt==1.3.0

#saya nambah catatan ini mas
django-autocomplete-light
django-user-agents
msoffcrypto-tool
# python3-memcached